Refactor: move MigrateToL2Information out of SignOrExecute
