+++ src/components/tx/confirmation-views/MigrateToL2Information/index.tsx
@@ -1,14 +1,21 @@
+import { useMemo } from 'react'
 import { Alert, AlertTitle, Box, SvgIcon, Typography } from '@mui/material'
+import type { TransactionData } from '@safe-global/safe-gateway-typescript-sdk'
 import InfoOutlinedIcon from '@/public/images/notifications/info.svg'
 import NamedAddressInfo from '@/components/common/NamedAddressInfo'
+import { extractMigrationL2MasterCopyAddress } from '@/utils/safe-migrations'
 
 export const MigrateToL2Information = ({
   variant,
-  newMasterCopy,
+  txData,
 }: {
   variant: 'history' | 'queue'
-  newMasterCopy?: string
+  txData?: TransactionData
 }) => {
+  const newMasterCopy = useMemo(() => {
+    return txData && extractMigrationL2MasterCopyAddress(txData)
+  }, [txData])
+
   return (
     <Box>
       <Alert severity="info" icon={<SvgIcon component={InfoOutlinedIcon} color="info" />}>
